deadline_project/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
django-crispy-forms
django-bootstrap-v5
python-decouple
